+++ cg/meta/orders/schema.py
@@ -17,6 +17,7 @@ class OrderType(StrEnum):
     BALSAMIC: str = str(Pipeline.BALSAMIC)
     EXTERNAL: str = "external"
     FASTQ: str = "fastq"
+    FLUFFY: str = str(Pipeline.FLUFFY)
     METAGENOME: str = "metagenome"
     MICROSALT: str = str(Pipeline.MICROSALT)
     MIP_DNA: str = str(Pipeline.MIP_DNA)
@@ -105,6 +106,7 @@ MIP_SAMPLE = {
     "source": OptionalNone(TypeValidatorNone(str)),
     "priority": OptionalNone(validators.Any(PRIORITY_OPTIONS)),
     "require_qcok": bool,
+    "volume": str,
     "container": OptionalNone(validators.Any(CONTAINER_OPTIONS)),
     # "required if plate for new samples"
     "container_name": OptionalNone(TypeValidatorNone(str)),
@@ -145,6 +147,7 @@ BALSAMIC_SAMPLE = {
     "sex": OptionalNone(validators.Any(SEX_OPTIONS)),
     "family_name": validators.RegexValidator(NAME_PATTERN),
     "require_qcok": bool,
+    "volume": str,
     "tumour": bool,
     "source": OptionalNone(TypeValidatorNone(str)),
     "priority": OptionalNone(validators.Any(PRIORITY_OPTIONS)),
@@ -177,6 +180,7 @@ MIP_RNA_SAMPLE = {
     "sex": OptionalNone(validators.Any(SEX_OPTIONS)),
     "source": OptionalNone(TypeValidatorNone(str)),
     "priority": OptionalNone(validators.Any(PRIORITY_OPTIONS)),
+    "volume": str,
     "container": OptionalNone(validators.Any(CONTAINER_OPTIONS)),
     # "required if plate for new samples"
     "container_name": OptionalNone(TypeValidatorNone(str)),
@@ -188,7 +192,7 @@ MIP_RNA_SAMPLE = {
     "quantity": OptionalNone(TypeValidatorNone(str)),
     "comment": OptionalNone(TypeValidatorNone(str)),
     # Orderform 1508:19
-    "from_sample": validators.RegexValidator(NAME_PATTERN),
+    "from_sample": OptionalNone(validators.RegexValidator(NAME_PATTERN)),
     "time_point": OptionalNone(TypeValidatorNone(str)),
 }
 
@@ -233,6 +237,7 @@ FASTQ_SAMPLE = {
     # todo: implement in OP or remove from OF
     # 'family_name': RegexValidator(NAME_PATTERN),
     "require_qcok": bool,
+    "volume": str,
     "source": str,
     "tumour": bool,
     "priority": OptionalNone(validators.Any(PRIORITY_OPTIONS)),