+++ cg/server/api.py
@@ -10,9 +10,9 @@ from google.auth import jwt
 from requests.exceptions import HTTPError
 from werkzeug.utils import secure_filename
 
-from cg.apps.lims import parse_json, parse_orderform
 from cg.constants import ANALYSIS_SOURCES, METAGENOME_SOURCES
-from cg.exc import DuplicateRecordError, OrderError, OrderFormError
+from cg.exc import DuplicateRecordError, OrderFormError, OrderError
+from cg.apps.lims import parse_orderform, parse_json_order
 from cg.meta.orders import OrdersAPI, OrderType
 
 from .ext import db, lims, osticket
@@ -106,14 +106,14 @@ def families():
         count = len(records)
     else:
         customer_obj = None if g.current_user.is_admin else g.current_user.customer
-        families_q = db.families(
+        case_query = db.families(
             enquiry=request.args.get("enquiry"),
             customer=customer_obj,
             action=request.args.get("action"),
         )
-        count = families_q.count()
-        records = families_q.limit(30)
-    data = [family_obj.to_dict(links=True) for family_obj in records]
+        count = case_query.count()
+        records = case_query.limit(30)
+    data = [case_obj.to_dict(links=True) for case_obj in records]
     return jsonify(families=data, total=count)
 
 
@@ -126,35 +126,35 @@ def families_in_customer_group():
     )
     count = families_q.count()
     records = families_q.limit(30)
-    data = [family_obj.to_dict(links=True) for family_obj in records]
+    data = [case_obj.to_dict(links=True) for case_obj in records]
     return jsonify(families=data, total=count)
 
 
 @BLUEPRINT.route("/families/<family_id>")
 def family(family_id):
     """Fetch a family with links."""
-    family_obj = db.family(family_id)
-    if family_obj is None:
+    case_obj = db.family(family_id)
+    if case_obj is None:
         return abort(404)
-    elif not g.current_user.is_admin and (g.current_user.customer != family_obj.customer):
+    if not g.current_user.is_admin and (g.current_user.customer != case_obj.customer):
         return abort(401)
 
-    data = family_obj.to_dict(links=True, analyses=True)
+    data = case_obj.to_dict(links=True, analyses=True)
     return jsonify(**data)
 
 
 @BLUEPRINT.route("/families_in_customer_group/<family_id>")
 def family_in_customer_group(family_id):
     """Fetch a family with links."""
-    family_obj = db.family(family_id)
-    if family_obj is None:
+    case_obj = db.family(family_id)
+    if case_obj is None:
         return abort(404)
     elif not g.current_user.is_admin and (
-        g.current_user.customer.customer_group != family_obj.customer.customer_group
+        g.current_user.customer.customer_group != case_obj.customer.customer_group
     ):
         return abort(401)
 
-    data = family_obj.to_dict(links=True, analyses=True)
+    data = case_obj.to_dict(links=True, analyses=True)
     return jsonify(**data)
 
 
@@ -339,11 +339,11 @@ def orderform():
             temp_dir = Path(tempfile.gettempdir())
             saved_path = str(temp_dir / filename)
             input_file.save(saved_path)
-            project_data = parse_orderform(saved_path)
+            parsed_order = parse_orderform(saved_path)
         else:
             json_data = json.load(input_file.stream)
-            project_data = parse_json(json_data)
+            parsed_order = parse_json_order(json_data)
     except OrderFormError as error:
         return abort(make_response(jsonify(message=error.message), 400))
 
-    return jsonify(**project_data)
+    return jsonify(**parsed_order)