+++ tests/cli/set/test_samples.py
@@ -53,9 +53,9 @@ def test_set_samples_by_invalid_identifier(
 
 def test_set_samples_by_case_id(cli_runner, base_context, base_store: Store, helpers, caplog):
     # GIVEN a database with a sample that belongs to a case
-    case_obj = helpers.add_family(store=base_store)
+    case_obj = helpers.add_case(store=base_store)
     sample_obj = helpers.add_sample(base_store)
-    helpers.add_relationship(store=base_store, family=case_obj, sample=sample_obj)
+    helpers.add_relationship(store=base_store, case=case_obj, sample=sample_obj)
 
     # WHEN calling set samples with case_id
     with caplog.at_level(logging.INFO):
@@ -93,9 +93,9 @@ def test_set_samples_by_valid_case_id_and_valid_identifier(
     cli_runner, base_context, base_store: Store, helpers, caplog
 ):
     # GIVEN a database with a sample that belongs to a case
-    case_obj = helpers.add_family(store=base_store)
+    case_obj = helpers.add_case(store=base_store)
     sample_obj = helpers.add_sample(base_store)
-    helpers.add_relationship(store=base_store, family=case_obj, sample=sample_obj)
+    helpers.add_relationship(store=base_store, case=case_obj, sample=sample_obj)
 
     # WHEN calling set samples with case_id for sample and valid identifier for sample
     with caplog.at_level(logging.INFO):
@@ -122,9 +122,9 @@ def test_set_samples_by_invalid_case_id_and_valid_identifier(
     cli_runner, base_context, base_store: Store, helpers, caplog
 ):
     # GIVEN a database with a sample that belongs to a case
-    case_obj = helpers.add_family(store=base_store)
+    case_obj = helpers.add_case(store=base_store)
     sample_obj = helpers.add_sample(base_store)
-    helpers.add_relationship(store=base_store, family=case_obj, sample=sample_obj)
+    helpers.add_relationship(store=base_store, case=case_obj, sample=sample_obj)
 
     # WHEN calling set samples with bad case_id for sample and valid identifier for sample
     with caplog.at_level(logging.INFO):
@@ -144,9 +144,9 @@ def test_set_samples_by_valid_case_id_and_invalid_identifier(
     cli_runner, base_context, base_store: Store, helpers, caplog
 ):
     # GIVEN a database with a sample that belongs to a case
-    case_obj = helpers.add_family(store=base_store)
+    case_obj = helpers.add_case(store=base_store)
     sample_obj = helpers.add_sample(base_store)
-    helpers.add_relationship(store=base_store, family=case_obj, sample=sample_obj)
+    helpers.add_relationship(store=base_store, case=case_obj, sample=sample_obj)
 
     # WHEN calling set samples with valid case_id for sample and wrong valid identifier for sample
     with caplog.at_level(logging.INFO):