+++ tests/meta/orders/test_meta_orders_api.py
@@ -36,9 +36,9 @@ def test_submit(base_store, orders_api: OrdersAPI, all_orders_to_submit, monkeyp
 
     # WHEN submitting the order
     order_ticket = {"name": "Paul Anderson", "email": "paul@magnolia.com"}
-    result = orders_api.submit(order_type, data=order_data, ticket=order_ticket)
+    result = orders_api.submit(order_type, order=order_data, ticket=order_ticket)
 
-    # THEN it should work...
+    # THEN the result should contain the ticket number for the order
     for record in result["records"]:
         if isinstance(record, models.Pool) or isinstance(record, models.Sample):
             assert record.ticket_number == ticket_number
@@ -88,7 +88,7 @@ def test_submit_illegal_sample_customer(
     # WHEN calling submit
     # THEN an OrderError should be raised on illegal customer
     with pytest.raises(OrderError):
-        orders_api.submit(order_type, data=order_data, ticket=order_ticket)
+        orders_api.submit(order_type, order=order_data, ticket=order_ticket)
 
 
 @pytest.mark.parametrize(
@@ -144,7 +144,7 @@ def test_submit_scout_legal_sample_customer(
 
     # WHEN calling submit
     # THEN an OrderError should not be raised on illegal customer
-    orders_api.submit(order_type, data=order_data, ticket=order_ticket)
+    orders_api.submit(order_type, order=order_data, ticket=order_ticket)
 
 
 @pytest.mark.parametrize(
@@ -201,4 +201,4 @@ def test_submit_non_scout_legal_sample_customer(
     # WHEN calling submit
     # THEN an OrderError should be raised on illegal customer
     with pytest.raises(OrderError):
-        orders_api.submit(order_type, data=order_data, ticket=order_ticket)
+        orders_api.submit(order_type, order=order_data, ticket=order_ticket)